requirements.txt
apscheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
python-telegram-bot==13.10
pytz-deprecation-shim==0.1.0.post0
pytz==2021.3
six==1.16.0
tornado==6.1
tzdata==2021.5; python_version >= '3.6'
tzlocal==4.1
